Update code generator and remove T4 templates
