+++ Source/QuantityTypes.sln
@@ -4,6 +4,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 VisualStudioVersion = 14.0.25420.1
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "QuantityTypes", "QuantityTypes\QuantityTypes.xproj", "{A6A7FA4A-2B1E-417E-B13C-50385ADB303C}"
+	ProjectSection(ProjectDependencies) = postProject
+		{D7AAF844-9614-4A44-991D-248DD447D624} = {D7AAF844-9614-4A44-991D-248DD447D624}
+	EndProjectSection
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CodeGenerator", "CodeGenerator\CodeGenerator.csproj", "{D7AAF844-9614-4A44-991D-248DD447D624}"
 EndProject
@@ -17,22 +20,6 @@ Global
 		{A6A7FA4A-2B1E-417E-B13C-50385ADB303C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A6A7FA4A-2B1E-417E-B13C-50385ADB303C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A6A7FA4A-2B1E-417E-B13C-50385ADB303C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{234E25BD-C3D0-465A-B763-69CCF5036CE9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{234E25BD-C3D0-465A-B763-69CCF5036CE9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{234E25BD-C3D0-465A-B763-69CCF5036CE9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{234E25BD-C3D0-465A-B763-69CCF5036CE9}.Release|Any CPU.Build.0 = Release|Any CPU
-		{C1EB77BC-D372-46E7-8A1D-8AF471CA7996}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C1EB77BC-D372-46E7-8A1D-8AF471CA7996}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C1EB77BC-D372-46E7-8A1D-8AF471CA7996}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C1EB77BC-D372-46E7-8A1D-8AF471CA7996}.Release|Any CPU.Build.0 = Release|Any CPU
-		{967118DC-80CF-4403-BAEA-0A23CC2D9B9B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{967118DC-80CF-4403-BAEA-0A23CC2D9B9B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{967118DC-80CF-4403-BAEA-0A23CC2D9B9B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{967118DC-80CF-4403-BAEA-0A23CC2D9B9B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E0557C29-6CDC-4060-BB1A-461ED5C1EF19}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E0557C29-6CDC-4060-BB1A-461ED5C1EF19}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E0557C29-6CDC-4060-BB1A-461ED5C1EF19}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E0557C29-6CDC-4060-BB1A-461ED5C1EF19}.Release|Any CPU.Build.0 = Release|Any CPU
 		{D7AAF844-9614-4A44-991D-248DD447D624}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{D7AAF844-9614-4A44-991D-248DD447D624}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D7AAF844-9614-4A44-991D-248DD447D624}.Release|Any CPU.ActiveCfg = Release|Any CPU
