--- Source/QuantityTypes/Quantities/CodeGenerator.cs
@@ -1,4 +1,0 @@
-﻿
-
-
-
